Remove unnecessary abstraction
The Mash database implementation was not Mash specific, so no need for
db interface.
