+++ src/us/kbase/assemblyhomology/minhash/MinHashSketchDatabase.java
@@ -1,23 +1,62 @@
 package us.kbase.assemblyhomology.minhash;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+import static us.kbase.assemblyhomology.util.Util.exceptOnEmpty;
+
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
-/** A database of MinHash sketches.
- * @author gaprice@lbl.gov
- *
- */
-public interface MinHashSketchDatabase {
-	
+import us.kbase.assemblyhomology.minhash.MinHashDBLocation;
+import us.kbase.assemblyhomology.minhash.MinHashImplementationInformation;
+import us.kbase.assemblyhomology.minhash.MinHashParameters;
+
+public class MinHashSketchDatabase {
+
+	//TODO TEST
 	//TODO JAVADOC
 
-	MinHashParameters getParameterSet();
+	private final MinHashImplementationInformation info;
+	private final MinHashParameters parameterSet;
+	private final MinHashDBLocation location;
+	private final List<String> sketchIDs;
 	
-	MinHashDBLocation getLocation();
-	
-	int getSketchCount();
-	
-	List<String> getSketchIDs();
+	public MinHashSketchDatabase(
+			final MinHashImplementationInformation info,
+			final MinHashParameters parameterSet,
+			final MinHashDBLocation location,
+			final List<String> sketchIDs) {
+		checkNotNull(info, "info");
+		checkNotNull(parameterSet, "parameterSet");
+		checkNotNull(location, "location");
+		checkNotNull(sketchIDs, "sketchIDs");
+		for (final String id: sketchIDs) {
+			exceptOnEmpty(id, "null or whitespace only id in sketchIDs");
+		}
+		this.info = info;
+		this.parameterSet = parameterSet;
+		this.location = location;
+		this.sketchIDs = Collections.unmodifiableList(new ArrayList<>(sketchIDs));
+	}
 
-	MinHashImplementationInformation getImplementationInformation(); 
+	public MinHashImplementationInformation getImplementationInformation() {
+		return info;
+	}
 	
+	public MinHashParameters getParameterSet() {
+		return parameterSet;
+	}
+
+	public MinHashDBLocation getLocation() {
+		return location;
+	}
+
+	public int getSketchCount() {
+		return sketchIDs.size();
+	}
+
+	public List<String> getSketchIDs() {
+		return sketchIDs;
+	}
+
 }
